requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
mysqlclient==2.0.3
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.2
